requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.4
Django==6.0.2
django-admin-interface==0.32.0
django-colorfield==0.14.0
django-rest-framework==0.1.0
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework-stubs==3.16.8
idna==3.11
pillow==12.1.1
psycopg2-binary==2.9.11
python-slugify==8.0.4
requests==2.32.5
sqlparse==0.5.5
text-unidecode==1.3
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
urllib3==2.6.3